remove options.h dependency on protobuf
